Move data connection parameters to its own component
